refactor(builder): extrahiere positive.rs aus single/mod.rs
- derive_positive_aspects_from_normalized → single/positive.rs
- Clone-Impls für FindingGroup, ExampleBlock, RoleAssignment → single/positive.rs
- normalized_module_score/grade bleiben in mod.rs (ziehen mit module_details.rs)
